pyproject.toml
[tool.poetry]
name = "<package-name>"
version = "0.1.0"
description = ""
authors = []
homepage = ""
readme = "README.md"
license = "MIT"
keywords = []
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.10",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
packages= [
    {include = "<package-name>", from = "src"},
]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
torch={version="2.1.2+cu121", source="torch_repo"}
torchvision={version="^0.17.1+cu121", source="torch_repo"}
hydra-core = "^1.3.2"
python-dotenv="^1"
wandb="^0.16.3"

[[tool.poetry.source]]
name = "torch_repo"
url = "https://download.pytorch.org/whl/cu121"
priority = "supplemental"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
black = "^24.2.0"
jupyter = "^1.0.0"
jupytext = "^1.16.1"

[tool.poetry.scripts]
main = "main:main"

[tool.poetry.urls]
"Issues" = "<github-repo>/issues"
"Repo" = "<github-repo>"


[build-system]
requires = ["poetry-core>=1.8.1"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 80
exclude ='''
/(
    artifacts/*
    | .venv/*
)
'''

[tool.pytest.ini_options]
pythonpath = [
    ".", "./src", "./tests"
]
addopts = "--cov src --cov-report html --cov-branch --import-mode importlib"
markers = [
    "integration",
    "slow",
    "dist",
]
norecursedirs = "tests/utils"
testpaths = "tests"
